# api/schema/kenaz/openapi.yaml
openapi: 3.1.0
info:
  description: Local-first knowledge base with Markdown storage, full-text search, and graph visualization.
  title: Kenaz API
  contact: {}
  version: 1.0.0
paths:
  /attachments:
    post:
      security:
        - BearerAuth: []
      tags:
        - attachments
      summary: Upload an attachment file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: File to upload
                  type: string
                  format: binary
              required:
                - file
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentUploadResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
  /graph:
    get:
      security:
        - BearerAuth: []
      tags:
        - graph
      summary: Get the knowledge graph
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphResponse"
  /notes:
    get:
      security:
        - BearerAuth: []
      tags:
        - notes
      summary: List notes with optional pagination and filtering
      parameters:
        - description: Page size
          name: limit
          in: query
          schema:
            type: integer
        - description: Page offset
          name: offset
          in: query
          schema:
            type: integer
        - description: Filter by tag
          name: tag
          in: query
          schema:
            type: string
        - description: Sort field
          name: sort
          in: query
          schema:
            type: string
            enum:
              - updated_at
              - title
              - path
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteListResponse"
    post:
      security:
        - BearerAuth: []
      tags:
        - notes
      summary: Create a new note
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
        description: Note to create
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDetail"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
  /notes/rename:
    post:
      security:
        - BearerAuth: []
      tags:
        - notes
      summary: Rename a note or directory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameNoteRequest"
        description: Old and new paths
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenameNoteResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
  /notes/{path}:
    get:
      security:
        - BearerAuth: []
      tags:
        - notes
      summary: Get a single note by path
      parameters:
        - description: Note path
          name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDetail"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
    put:
      security:
        - BearerAuth: []
      tags:
        - notes
      summary: Update a note with optimistic concurrency
      parameters:
        - description: Note path
          name: path
          in: path
          required: true
          schema:
            type: string
        - description: SHA-256 checksum for optimistic concurrency
          name: If-Match
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
        description: Updated content
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDetail"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
    delete:
      security:
        - BearerAuth: []
      tags:
        - notes
      summary: Delete a note
      parameters:
        - description: Note path
          name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Note deleted
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/errResponse"
  /search:
    get:
      security:
        - BearerAuth: []
      tags:
        - search
      summary: Full-text search across notes
      parameters:
        - description: Search query
          name: q
          in: query
          required: true
          schema:
            type: string
        - description: Max results
          name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errResponse"
servers:
  - url: /api
    description: Default (relative)
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
    AttachmentUploadResponse:
      type: object
      required:
        - filename
        - size
        - url
      properties:
        filename:
          type: string
          example: image.png
        size:
          type: integer
          example: 12345
        url:
          type: string
          example: /attachments/image.png
    CreateNoteRequest:
      type: object
      required:
        - content
        - path
      properties:
        content:
          type: string
          example: |-
            # Hello
            World
        path:
          type: string
          example: notes/hello.md
    GraphLink:
      type: object
      required:
        - source
        - target
      properties:
        source:
          type: string
          example: notes/hello.md
        target:
          type: string
          example: notes/world.md
    GraphNode:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          example: notes/hello.md
        title:
          type: string
          example: Hello
    GraphResponse:
      type: object
      required:
        - links
        - nodes
      properties:
        links:
          type: array
          items:
            $ref: "#/components/schemas/GraphLink"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
    NoteDetail:
      type: object
      required:
        - backlinks
        - checksum
        - content
        - path
        - tags
        - title
        - updated_at
      properties:
        backlinks:
          type: array
          items:
            type: string
        checksum:
          type: string
        content:
          type: string
        frontmatter:
          type: object
          additionalProperties: {}
        path:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        updated_at:
          type: string
    NoteListItem:
      type: object
      required:
        - checksum
        - path
        - tags
        - title
        - updated_at
      properties:
        checksum:
          type: string
        path:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        updated_at:
          type: string
    NoteListResponse:
      type: object
      required:
        - notes
        - total
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/NoteListItem"
        total:
          type: integer
          example: 42
    RenameNoteRequest:
      type: object
      required:
        - new_path
        - old_path
      properties:
        new_path:
          type: string
          example: notes/new.md
        old_path:
          type: string
          example: notes/old.md
    RenameNoteResponse:
      type: object
      properties:
        backlinks:
          type: array
          items:
            type: string
        checksum:
          type: string
        content:
          type: string
        frontmatter:
          type: object
          additionalProperties: {}
        moved:
          description: Directory rename field.
          type: array
          items:
            type: string
        path:
          description: Note rename fields (same as NoteDetail).
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        updated_at:
          type: string
    SearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
    SearchResult:
      type: object
      required:
        - path
        - snippet
        - title
      properties:
        path:
          type: string
          example: notes/hello.md
        snippet:
          type: string
          example: ...matched text...
        title:
          type: string
          example: Hello
    UpdateNoteRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: |-
            # Updated
            Content
    errResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string